requirements.txt
openai==1.*
duckduckgo-search==6.*
trafilatura==1.*
pypdf==5.*
tavily-python==0.*
python-dotenv==1.*
requests==2.*
aiolimiter==1.*
